chatbot/requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-groq==0.1.9
langchain-huggingface==0.0.3
langchain-chroma
chromadb
uvicorn
fastapi
python-dotenv